Trait fs_node_runtime::BlockT

pub trait BlockT: 'static + Clone + Send + Sync + Codec + Eq + MaybeSerialize + Debug + MaybeMallocSizeOf {
    type Extrinsic: Member + Codec + Extrinsic + MaybeSerialize + MaybeMallocSizeOf;
    type Header: Header + MaybeMallocSizeOf
   where
        <Self::Header as Header>::Hash == Self::Hash;
    type Hash: Member + MaybeSerializeDeserialize + Debug + Hash + Ord + Copy + MaybeDisplay + Default + SimpleBitOps + Codec + AsRef<[u8]> + AsMut<[u8]> + MaybeMallocSizeOf + TypeInfo;

    fn header(&self) -> &Self::Header;
    fn extrinsics(&self) -> &[Self::Extrinsic];
    fn deconstruct(self) -> (Self::Header, Vec<Self::Extrinsic, Global>);
    fn new(
        header: Self::Header,
        extrinsics: Vec<Self::Extrinsic, Global>
    ) -> Self;
    fn encode_from(
        header: &Self::Header,
        extrinsics: &[Self::Extrinsic]
    ) -> Vec<u8, Global>;

    fn hash(&self) -> Self::Hash { ... }
}

Something which fulfills the abstract idea of a Substrate block. It has types for Extrinsic pieces of information as well as a Header.

You can get an iterator over each of the extrinsics and retrieve the header.

Required Associated Types

type Extrinsic: Member + Codec + Extrinsic + MaybeSerialize + MaybeMallocSizeOf

Type for extrinsics.

type Header: Header + MaybeMallocSizeOf
where
<Self::Header as Header>::Hash == Self::Hash

Header type.

type Hash: Member + MaybeSerializeDeserialize + Debug + Hash + Ord + Copy + MaybeDisplay + Default + SimpleBitOps + Codec + AsRef<[u8]> + AsMut<[u8]> + MaybeMallocSizeOf + TypeInfo

Block hash type.

Required Methods

fn header(&self) -> &Self::Header

Returns a reference to the header.

fn extrinsics(&self) -> &[Self::Extrinsic]

Returns a reference to the list of extrinsics.

fn deconstruct(self) -> (Self::Header, Vec<Self::Extrinsic, Global>)

Split the block into header and list of extrinsics.

fn new(header: Self::Header, extrinsics: Vec<Self::Extrinsic, Global>) -> Self

Creates new block from header and extrinsics.

fn encode_from(
header: &Self::Header,
extrinsics: &[Self::Extrinsic]
) -> Vec<u8, Global>

Creates an encoded block from the given header and extrinsics without requiring the creation of an instance.

Provided Methods

fn hash(&self) -> Self::Hash

Returns the hash of the block.